% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeq_mort.R
\name{aeq_mort}
\alias{aeq_mort}
\title{Extract AEQ mortality from Chinook FRAM database}
\usage{
aeq_mort(db, runs = NULL, stocks = NULL, drop_t1 = T, sum_ages = F)
}
\arguments{
\item{db}{string, file path to database}

\item{runs}{numeric, RunID(s) as ID or ID:ID}

\item{stocks}{numeric, StockID(s) as ID or ID:ID}

\item{drop_t1}{logical, should timestep 1 be excluded?}

\item{sum_ages}{logical, should ages 2:5 be summed? (per run-s-f-t)}
}
\value{
a tibble of AEQ'd mortality, possibly aggregated over ages to
the per run-stock-fishery-timestep total
}
\description{
Extract AEQ mortality from Chinook FRAM database
}
\examples{
\dontrun{

#showing M&UM for a given stock across runs
#with ages 2:5 summed for timesteps 2:4
aeq_mort("path/to/Chinook_FRAM_Database.mdb",
 runs = 1:4, stocks = 31:32, drop_t1 = T, sum_ages = T)

}
}
