% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayock_split.R
\name{ayock_split}
\alias{ayock_split}
\title{Convenience function to perform the "Ayock Splits"}
\usage{
ayock_split(db, tamm_file_paths)
}
\arguments{
\item{db}{string, file path to database}

\item{tamm_file_paths}{character vector of path(s)
to TAMM xlsx associated with \code{runs}}
}
\value{
Nothing, but xlsx files outside of R/Rstudio should be altered,
and the metadata and values are printed to the console.
}
\description{
This function extracts and summarizes the Chin FRAM
Mortality table info for HC FF via \code{framr::aeq_mort}.

The timestep 3 AEQ'd NS and MSF mortality for HC FF
(UM:31 and M:32, including 2s) in A12 sport are then written into
the appropriate TAMM cells via the non-CRAN package \code{RDCOMClient}.
This dependency can be installed with:

install.packages("RDCOMClient", repos = "http://www.omegahat.net/R")
}
