% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-chin_stock_lu.R
\docType{data}
\name{chin_stock_lu}
\alias{chin_stock_lu}
\title{Lookup of Chinook FRAM stocks.}
\format{
data frame (tbl_df) with 116 rows and 4 cols:
\describe{
\item{bkfram_id}{numeric BKFRAM ID}
\item{StockID}{numeric forward FRAM stock ID}
\item{StockName}{character forward FRAM stock name}
\item{StockLongName}{character forward FRAM long name}
}
}
\source{
R7.1 Chinook FRAM project database tables 'Stock'
and 'BackwardsFRAM', with derivation in chin_valid_2020.Rmd
from FRAMEscapeV2 sheets in ChinRSScalers and Valid2018 workbooks
}
\usage{
chin_stock_lu
}
\description{
Containing BackwardsFRAM and 'forward' StockIDs.
The BackwardsFRAM table contains additional rows
of summed marked and unmarked per-stock; these are
NA for forward fields.
}
\examples{
chin_stock_lu |> dplyr::filter(!is.na(StockID))
}
\keyword{datasets}
