% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_coho_fish_sclr.R
\name{load_coho_fish_sclr}
\alias{load_coho_fish_sclr}
\title{Load FisheryScaler values into a Coho FRAM database}
\usage{
load_coho_fish_sclr(db, runs = NULL, fs_new = NULL)
}
\arguments{
\item{db}{string, file path to database}

\item{runs}{numeric, RunID(s) as ID or ID:ID, default NULL for all runs}

\item{fs_new}{tibble/data.frame, object with FisheryID, TimeStep and Flag/Scaler/Quota fields}
}
\value{
nothing
}
\description{
INCOMPLETE DO NOT YET ATTEMPT TO USE
}
