% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coho_escp.R
\name{read_coho_escp}
\alias{read_coho_escp}
\title{Extract escapement values from Coho FRAM database}
\usage{
read_coho_escp(db, runs = NULL, stocks = NULL)
}
\arguments{
\item{db}{string, file path to database}

\item{runs}{numeric, RunID(s) as ID or ID:ID, default NULL for all runs}

\item{stocks}{numeric, StockID(s) as ID or ID:ID, default NULL for all stocks}
}
\value{
a tibble of escapement estimates; already only age 3 in TimeStep 5
}
\description{
Extract escapement values from Coho FRAM database
}
\examples{
\dontrun{

#fetch all runs for WA stocks
read_coho_escp("path/to/FRAM_Database.mdb", stocks = 1:164)
#fetch a single run for a single stock
read_coho_escp("path/to/FRAM_Database.mdb", stocks = 161)

}

}
