% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coho_fish_sclr.R
\name{read_coho_fish_sclr}
\alias{read_coho_fish_sclr}
\title{Extract FisheryScaler values from Coho FRAM database}
\usage{
read_coho_fish_sclr(db, runs = NULL, fisheries = NULL)
}
\arguments{
\item{db}{string, file path to database}

\item{runs}{numeric, RunID(s) as ID or ID:ID, default NULL for all runs}

\item{fisheries}{numeric, FisheryID(s) as ID or ID:ID, default NULL for all fisheries}
}
\value{
a tibble of non-selective and mark-selective input
fishery catch estimates, with mark-selective parameters where applicable
}
\description{
Extract FisheryScaler values from Coho FRAM database
}
