% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coho_mort.R
\name{read_coho_mort}
\alias{read_coho_mort}
\title{Extract mortality values from Coho FRAM database}
\usage{
read_coho_mort(db, runs = NULL, stocks = NULL, fisheries = NULL)
}
\arguments{
\item{db}{string, file path to database}

\item{runs}{numeric, RunID(s) as ID or ID:ID, default NULL for all runs}

\item{stocks}{numeric, StockID(s) as ID or ID:ID, default NULL for all stocks}

\item{fisheries}{numeric, FisheryID(s) as ID or ID:ID, default NULL for all fisheries}
}
\value{
a tibble of mortality by run-stock-fishery-timestep
}
\description{
Extract mortality values from Coho FRAM database
}
\examples{
\dontrun{

#fetch all runs for WA stocks across all fisheries
read_coho_mort("path/to/FRAM_Database.mdb", stocks = 1:164)
#fetch a single stock in a single fishery across runs/years
read_coho_mort("path/to/FRAM_Database.mdb", stocks = 161, fisheries = 45)

}

}
