% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_coho_tamm_nttr.R
\name{read_coho_tamm_nttr}
\alias{read_coho_tamm_nttr}
\title{Extract values from coho TAMM sheet NT-Tsummary}
\usage{
read_coho_tamm_nttr(tamm)
}
\arguments{
\item{tamm}{string, file path to TAMM workbook}
}
\value{
a of mortality and ER values by fishery and stock regions,
for "hat", "nat" (labeled "wild" in TAMM) and "tot" groupings.
NOTE the formulas in this sheet for "US OCEAN" include Alaskan fisheries,
and the "all" columns for these fisheries include non-WA stocks;
these values are typically hidden in sheet "2". See example below
for an exclude filter example.
}
\description{
Extract values from coho TAMM sheet NT-Tsummary
}
\examples{
\dontrun{

read_coho_tamm_nttr("path/to/TAMM/file.xlsx") |>
 #dropping queets_sup and "all_um/all_m" cols that include non-WA stocks in the AK catches
 filter(!is.na(origin), stk_type != "all_tot")

}

}
