% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_ocean_opt.R
\name{rep_ocean_opt}
\alias{rep_ocean_opt}
\title{Convenience function to replicate PFMC fishery inputs across FRAM RunIDs}
\usage{
rep_ocean_opt(db, run_from, run_to)
}
\arguments{
\item{db}{string, file path to database}

\item{run_from}{numeric, "donor" run}

\item{run_to}{numeric, "recipient" run}
}
\value{
Nothing, but database tables are altered.
}
\description{
This updates a OceanOptMergeFile.R script
developed by Jon Carey. Maintains DELETE then append pattern,
avoiding possibly slow rewrite of entire FisheryScalers
in a project mdb with many runs. Also left as single pairwise RunIDs
and fixed FisheryIDs rather than passing as args.
}
\examples{

}
