% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_ocean_opt_coho.R
\name{rep_ocean_opt_coho}
\alias{rep_ocean_opt_coho}
\title{Convenience function to replicate fishery inputs across FRAM RunIDs}
\usage{
rep_ocean_opt_coho(db, run_from, run_to)
}
\arguments{
\item{db}{string, file path to database}

\item{run_from}{numeric, "donor" run}

\item{run_to}{numeric, "recipient" run}
}
\value{
Nothing, but database tables are altered.
}
\description{
This transfers FisheryScalers and NonRetention table
values from one to another run, as is necessary during PFMC STT
examination of multiple ocean options.
Maintains DELETE then append pattern originally developed by Jon Carey,
avoiding possibly slow rewrite of entire FisheryScalers
in a project mdb with many runs.
}
