% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps.R
\name{sps}
\alias{sps}
\title{Convenience function to perform in-workbook SPS iterations}
\usage{
sps(tamm_file_paths)
}
\arguments{
\item{tamm_file_paths}{character vector of path(s)}
}
\value{
Nothing, but xlsx files outside of R/Rstudio should be altered
}
\description{
This function iterates the copy/paste process
in a Chinook TAMM, leaving the actual calculations to the Excel
formulas. It is fundamentally unchanged from older work by
Jon Carey (NOAA) and Derek Dapp (WDFW).
}
