% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_2s.R
\name{update_2s}
\alias{update_2s}
\title{Convenience function to update age 2 recruit scalers given forecast 2s}
\usage{
update_2s(chinrs, db, runID)
}
\arguments{
\item{chinrs}{string, path to "ChinRSScalers" file with provided a2s}

\item{db}{string, file path to database}

\item{runID}{numeric, RunID to be altered}
}
\value{
Nothing, but database tables are altered.
}
\description{
This function modernizes scripts developed by Jon Carey
and Derek Dapp to update and overwrite Chinook age 2 recruit scalers
in a database run. It calculates the ratio of the provided forecast
age 2 abundance to the existing mature cohort abundance, and then
multiplies that ratio against the existing recruit scaler.
}
\examples{
\dontrun{
#note that reads from xlsm ChinRSScaler files can be very slow
#due to old/leftover sheet names & connections
update_2s("path/to/ChinRSScalarsMR21.xlsm", "path/to/Chinook_FRAM_Database.mdb", runID = NNN)

}
}
